pyproject.toml
[project]
name = "nexusflow"
version = "0.1.0"
description = "NexusFlow - Intelligent Ticket Auto-Resolution and Classification System"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "NexusFlow Team" }]

dependencies = [
    # Core Framework
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    
    # LangGraph and LangChain (Latest v1.x versions)
    "langgraph>=1.0.5",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langchain-anthropic>=1.3.1",
    "langchain-community>=1.2.0",
    "langchain-core>=1.2.6",
    
    # MCP Server (FastMCP 2.x)
    "fastmcp>=2.14.2",
    
    # Databases
    "neo4j>=5.28.0",
    "pymilvus>=2.5.1",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy[asyncio]>=2.0.37",
    
    # OpenAI
    "openai>=1.60.0",
    "tiktoken>=0.8.0",
    
    # Anthropic
    "anthropic>=0.43.0",
    
    # Observability (Arize Phoenix)
    "arize-phoenix>=6.1.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "openinference-instrumentation-langchain>=0.1.31",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    
    # Data Processing
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    
    # Utilities
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "rich>=13.9.4",
    "typer>=0.15.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
]

[project.scripts]
nexusflow = "nexusflow.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexusflow"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
